feat: 🎸 px board update

+++ web3/pixel-board/contract/src/shared.rs
@@ -1,2 +1,2 @@
-pub type Position = (usize, usize);
+pub type Position = (u32, u32);
 pub type Color = (u8, u8, u8, u8);
